add postinstall question to simple questions
